Refactor: Relocate `TrackingAnnotationInfo` and remove `AnnotationScanner`
This commit refactors the plugin's utility classes by moving `TrackingAnnotationInfo` to its own file and removing the now-unused `AnnotationScanner`.

**Key changes:**

*   **New `TrackingAnnotationInfo.kt`**:
    *   The `TrackingAnnotationInfo` data class and its nested `AnnotationType` enum have been moved from `AnnotationScanner.kt` to a new dedicated file, `plugin/src/main/kotlin/com/shalan/analytics/plugin/utils/TrackingAnnotationInfo.kt`.
    *   KDoc documentation has been added to the class and its properties to improve clarity and maintainability.

*   **Removed `AnnotationScanner.kt`**:
    *   The `AnnotationScanner.kt` file, which contained the logic for scanning class files for tracking annotations, has been deleted.
    *   This component is no longer needed as the new instrumentation approach in `AnalyticsClassVisitorFactory` handles annotation detection directly during the class visit, making a separate scanning phase redundant.

*   **Removed `AnnotationScannerTest.kt`**:
    *   The corresponding test file for the `AnnotationScanner` has also been deleted.

This change streamlines the plugin's internal structure by removing an unnecessary layer of abstraction and co-locating the data class in its own file, following best practices.
